Move properties to a separate module
