+++ src/domain/property/mod.rs
@@ -1,0 +1,5 @@
+mod boolean_property;
+mod property;
+
+pub use boolean_property::BooleanProperty;
+pub use property::{Property, PropertyType};
